refactor: QR Code Dialog
