+++ packages/mask/shared-ui/locales/en-US.json
@@ -397,12 +397,10 @@
     "plugin_wallet_pending_transactions_other": "{{count}} Pendings",
     "plugin_wallet_import_wallet": "Import Wallet",
     "plugin_wallet_select_provider_dialog_title": "Connect Wallet",
-    "plugin_wallet_qr_code_with_wallet_connect": "Scan QR code with a WalletConnect-compatible wallet",
     "plugin_wallet_token_unlock": "Exact Unlock",
     "plugin_wallet_token_infinite_unlock": "Unlock {{symbol}}",
     "plugin_wallet_token_exact_unlock_tips": "You must give the {{provider}} smart contract permission to use your part of {{symbol}}. ",
     "plugin_wallet_token_infinite_unlock_tips": "You must give the {{provider}} smart contract permission to use your {{symbol}}. You only have to do this once per token.",
-    "plugin_wallet_connect_dialog_title": "WalletConnect",
     "plugin_wallet_connect_safari_metamask": "Connect to MetaMask",
     "plugin_wallet_connect_safari_rainbow": "Connect to Rainbow",
     "plugin_wallet_connect_safari_trust": "Connect to Trust",
@@ -414,7 +412,6 @@
     "plugin_wallet_switch_network_under_going": "Switching to {{network}}",
     "plugin_wallet_connect_tips": "Please switch to this network in the mobile\napplication wallet you are connected to.",
     "plugin_wallet_not_support_network": "This network is not supported yet.",
-
     "plugin_wallet_connect_network": "Change to {{network}}",
     "plugin_wallet_not_available_on": "Not available on {{network}} Network.",
     "plugin_wallet_connect_wallet": "Connect Wallet",
