+++ packages/mask/shared-ui/locales/ko-KR.json
@@ -365,12 +365,10 @@
     "plugin_wallet_pending_transactions_other": "{{count}} 대기 중",
     "plugin_wallet_import_wallet": "월렛 불러오기",
     "plugin_wallet_select_provider_dialog_title": "월렛 연결",
-    "plugin_wallet_qr_code_with_wallet_connect": "WalletConnect와 호환성이 있는 월렛으로 QR 코드를 스칸하세요.",
     "plugin_wallet_token_unlock": "정확 언락",
     "plugin_wallet_token_infinite_unlock": "무한 언락",
     "plugin_wallet_token_exact_unlock_tips": "{{provider}} 에게 {{symbol}} 이용할 수 있는 스마트 컨트랙트 권한을 줘야 합니다. ",
     "plugin_wallet_token_infinite_unlock_tips": "{{provider}} 에게 {{symbol}} 이용할 수 있는 스마트 컨트랙트 권한을 줘야 합니다. 다른 토큰을 이용할 때 이 작업을 다시 해야 합니다.",
-    "plugin_wallet_connect_dialog_title": "월렛 연결",
     "plugin_wallet_connect_safari_metamask": "MetaMask 연결하기",
     "plugin_wallet_connect_safari_rainbow": "Rainbow 연결하기",
     "plugin_wallet_connect_safari_trust": "Trust 연결하기",
