+++ packages/mask/shared-ui/locales/qya-AA.json
@@ -397,12 +397,10 @@
     "plugin_wallet_pending_transactions_other": "crwdns18186:0{{count}}crwdne18186:0",
     "plugin_wallet_import_wallet": "crwdns4641:0crwdne4641:0",
     "plugin_wallet_select_provider_dialog_title": "crwdns4663:0crwdne4663:0",
-    "plugin_wallet_qr_code_with_wallet_connect": "crwdns4665:0crwdne4665:0",
     "plugin_wallet_token_unlock": "crwdns4667:0crwdne4667:0",
     "plugin_wallet_token_infinite_unlock": "crwdns4669:0{{symbol}}crwdne4669:0",
     "plugin_wallet_token_exact_unlock_tips": "crwdns17548:0{{provider}}crwdnd17548:0{{symbol}}crwdne17548:0",
     "plugin_wallet_token_infinite_unlock_tips": "crwdns17550:0{{provider}}crwdnd17550:0{{symbol}}crwdne17550:0",
-    "plugin_wallet_connect_dialog_title": "crwdns4671:0crwdne4671:0",
     "plugin_wallet_connect_safari_metamask": "crwdns4673:0crwdne4673:0",
     "plugin_wallet_connect_safari_rainbow": "crwdns4675:0crwdne4675:0",
     "plugin_wallet_connect_safari_trust": "crwdns4677:0crwdne4677:0",
