+++ packages/mask/shared-ui/locales/zh-CN.json
@@ -395,12 +395,10 @@
     "plugin_wallet_pending_transactions_other": "{{count}} 个待定",
     "plugin_wallet_import_wallet": "导入钱包",
     "plugin_wallet_select_provider_dialog_title": "连接钱包",
-    "plugin_wallet_qr_code_with_wallet_connect": "使用WalletConnect兼容的钱包扫描二维码",
     "plugin_wallet_token_unlock": "精确解锁",
     "plugin_wallet_token_infinite_unlock": "无限解锁",
     "plugin_wallet_token_exact_unlock_tips": "您必须授予 {{provider}} 智能合约权限才能使用您的 {{symbol}}。 ",
     "plugin_wallet_token_infinite_unlock_tips": "您必须授予 {{provider}} 智能合约权限才能使用您的 {{symbol}} 。每个代币只需要授权一次。",
-    "plugin_wallet_connect_dialog_title": "WalletConnect",
     "plugin_wallet_connect_safari_metamask": "连接到 MetaMask",
     "plugin_wallet_connect_safari_rainbow": "连接到 Rainbow",
     "plugin_wallet_connect_safari_trust": "连接到 Trust",
