+++ packages/mask/shared-ui/locales/zh-TW.json
@@ -169,10 +169,8 @@
     "plugin_wallet_on_connect": "連接錢包",
     "plugin_wallet_import_wallet": "匯入錢包",
     "plugin_wallet_select_provider_dialog_title": "連接錢包",
-    "plugin_wallet_qr_code_with_wallet_connect": "使用一個與 WalletConnect 相容的錢包掃描 QR code",
     "plugin_wallet_token_unlock": "特定數量解鎖",
     "plugin_wallet_token_infinite_unlock": "無限解鎖",
-    "plugin_wallet_connect_dialog_title": "WalletConnect",
     "plugin_wallet_connect_safari_metamask": "連接到 MetaMask",
     "plugin_wallet_connect_safari_rainbow": "連接到 Rainbow",
     "plugin_wallet_connect_safari_trust": "連接到 Trust",
