+++ packages/mask/src/plugins/Wallet/Dashboard/index.tsx
@@ -8,7 +8,6 @@ import { ConnectWalletDialog } from '../SNSAdaptor/ConnectWalletDialog/index.js'
 import { WalletRiskWarningDialog } from '../SNSAdaptor/RiskWarningDialog/index.js'
 import { GasSettingDialog } from '../SNSAdaptor/GasSettingDialog/index.js'
 import { TransactionSnackbar } from '../SNSAdaptor/TransactionSnackbar/index.js'
-import { WalletConnectQRCodeDialog } from '../SNSAdaptor/WalletConnectQRCodeDialog/index.js'
 
 const dashboard: Plugin.Dashboard.Definition = {
     ...base,
@@ -25,7 +24,6 @@ const dashboard: Plugin.Dashboard.Definition = {
                 {getSiteType() !== ExtensionSite.Popup ? (
                     <TransactionSnackbar pluginID={NetworkPluginID.PLUGIN_EVM} />
                 ) : null}
-                <WalletConnectQRCodeDialog />
             </>
         )
     },
