+++ packages/shared/src/locales/en-US.json
@@ -146,8 +146,8 @@
     "plugin_wallet_not_support_network": "This network is not supported yet.",
     "plugin_wallet_switch_network_under_going": "Switching to {{network}}",
     "plugin_wallet_connect_tips": "Please switch to this network in the mobile\napplication wallet you are connected to.",
-    "plugin_wallet_not_available_on": "Not available on {{network}} Network.",
     "plugin_wallet_connect_network": "Change to {{network}}",
+    "plugin_wallet_not_available_on": "Not available on {{network}} Network.",
     "plugin_wallet_token_infinite_unlock_tips": "You must give the {{provider}} smart contract permission to use your {{symbol}}. You only have to do this once per token.",
     "plugin_wallet_token_infinite_unlock": "Unlock {{symbol}}",
     "plugin_wallet_select_a_nft_contract": "Select an NFT Contract",
@@ -158,6 +158,8 @@
     "plugin_wallet_approve_all_nft_successfully": "Unlock {{symbol}} successfully",
     "plugin_wallet_fail_to_load_nft_contract": "Click to retry",
     "plugin_wallet_change_wallet": "Change Wallet",
+    "wallet_connect_qr_code_dialog_title": "WalletConnect",
+    "wallet_connect_qr_code_dialog_content": "Scan QR code with a WalletConnect-compatible wallet",
     "save": "Save",
     "cancel": "Cancel",
     "add_wallet": "Add Wallet",
