+++ packages/shared/src/locales/ko-KR.json
@@ -3,5 +3,7 @@
     "import": "불러오기",
     "balance": "잔액",
     "select_token": "토큰 선택",
-    "erc20_token_list_placeholder": "토큰 이름이나 컨트렉트 주소"
+    "erc20_token_list_placeholder": "토큰 이름이나 컨트렉트 주소",
+    "wallet_connect_qr_code_dialog_title": "월렛 연결",
+    "wallet_connect_qr_code_dialog_content": "WalletConnect와 호환성이 있는 월렛으로 QR 코드를 스칸하세요."
 }
