+++ packages/web3-modals/src/modals/WalletConnectQRCodeDialog/QRCodeModal.tsx
@@ -1,7 +1,6 @@
 import { Grid, Typography } from '@mui/material'
 import { makeStyles } from '@masknet/theme'
-import { QRCode } from '@masknet/shared'
-import { useI18N } from '../../../../utils/index.js'
+import { QRCode, useSharedI18N } from '@masknet/shared'
 
 const useStyles = makeStyles()((theme) => ({
     container: {
@@ -18,7 +17,7 @@ const useStyles = makeStyles()((theme) => ({
 export const QRCodeModal: React.FC<{
     uri: string
 }> = ({ uri }) => {
-    const { t } = useI18N()
+    const t = useSharedI18N()
 
     const { classes } = useStyles()
     const style: React.CSSProperties = {
@@ -32,7 +31,7 @@ export const QRCodeModal: React.FC<{
         <Grid className={classes.container}>
             <QRCode text={uri} canvasProps={{ style }} />
             <Typography className={classes.tip} color="textSecondary">
-                {t('plugin_wallet_qr_code_with_wallet_connect')}
+                {t.wallet_connect_qr_code_dialog_content()}
             </Typography>
         </Grid>
     )
